requirements.txt
acres==0.2.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.1.0
dataclasses-json==0.6.7
etelemetry==0.3.1
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
fitz==0.0.1.dev2
frozenlist==1.5.0
fsspec==2025.2.0
gradio==5.15.0
gradio_client==1.7.0
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.10
isodate==0.6.1
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-milvus==0.1.8
langchain-ollama==0.2.3
langchain-text-splitters==0.3.5
langsmith==0.3.4
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.2
ollama==0.4.7
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.1.0
propcache==0.2.1
protobuf==5.29.3
prov==2.0.1
puremagic==1.28
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydot==3.0.4
pydub==0.25.1
Pygments==2.19.1
pymilvus==2.5.4
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
ruff==0.9.4
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
semantic-version==2.10.0
sentence-transformers==3.4.1
setuptools==75.8.0
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.6.0
tqdm==4.67.1
traits==7.0.2
transformers==4.48.2
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
yarl==1.18.3
zstandard==0.23.0
